Migrate quote parsing to IMDb quotes
Quotes from rottentomatoes had a number of issues including duplicates and a massive amount of different formatting. IMDb quotes appear to be more curated and more uniform in formatting.

IMDb quotes do have a number of differing formats that required detection and corrections. This commit includes detection and correction for a few of the quote formats. However, there are other formats that detection and correction needs to be added later.

The quote duplicate check does not appear to be needed for IMDb but the code is left commented until further testing with IMDb.